Safari-611.1.10/LayoutTests/accessibility/mac/landmark-not-ignored-expected.txt

header
aside
main
nav
article
section
This tests that HTML5 elements which have implicit landmark and related semantics are not ignored if their render objects have block render objects children.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS test.childrenCount is 6
PASS el.role is "AXRole: AXGroup"
PASS el.subrole is "AXSubrole: AXLandmarkBanner"
PASS el.role is "AXRole: AXGroup"
PASS el.subrole is "AXSubrole: AXLandmarkComplementary"
PASS el.role is "AXRole: AXGroup"
PASS el.subrole is "AXSubrole: AXLandmarkMain"
PASS el.role is "AXRole: AXGroup"
PASS el.subrole is "AXSubrole: AXLandmarkNavigation"
PASS el.role is "AXRole: AXGroup"
PASS el.subrole is "AXSubrole: AXDocumentArticle"
PASS el.role is "AXRole: AXGroup"
PASS el.subrole is "AXSubrole: "
PASS successfullyParsed is true

TEST COMPLETE
